# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for EC-Wallet application",
        "title": "EC-Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/wallet/v1/payment-address": {
            "get": {
                "description": "Allocate a BNB address for a specific order number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Get payment address for an order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallet.PaymentAddressResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid order ID",
                        "schema": {
                            "$ref": "#/definitions/wallet.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/wallet.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "wallet.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "Invalid order ID"
                }
            }
        },
        "wallet.PaymentAddressResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "bnb1w7jflwesfnrp0lfgnthkvq55m8gzrlav5ktmyk"
                },
                "chain": {
                    "type": "string",
                    "example": "BNB"
                },
                "created_at": {
                    "type": "string"
                },
                "expire_time": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string",
                    "example": "ORD12345678"
                }
            }
        }
    }
}